+++ src/scripts.go
@@ -32,6 +32,7 @@ apt-get -y install libcurl4 || exit 1
 apt-get -y install ldap-utils || exit 1
 apt-get -y install python3-setuptools || exit 1
 apt-get -y install iptables wget dnsutils tcpdump python net-tools vim binutils iproute2 python3 libcurl4-openssl-dev less || exit 1
+apt-get -y install dnsutils iputils-ping telnet netcat sysstat
 wget https://github.com/bestmethod/tcconfig-ubuntu-venv/archive/tc.tar.gz || exit 1
 tar -zxvf tc.tar.gz || exit 1
 mv tcconfig-ubuntu-venv-tc /tcconfig || exit 1
@@ -80,6 +81,7 @@ yum -y install dnsutils || yum -y install bind-utils
 yum -y install python
 yum -y install initscripts || exit 1
 yum -y install redhat-lsb || exit 1
+yum -y install telnet sysstat nc bind-utils iputils
 yum -y install centos-release-scl ; yum install -y rh-python36 ; yum -y install python38 || yum -y install python36
 cd /root && tar -zxvf installer.tgz || exit 1
 cd aerospike-server-* && ./asinstall || exit 1
@@ -112,6 +114,7 @@ yum -y install dnsutils || yum -y install bind-utils
 yum -y install python
 yum -y install initscripts || exit 1
 yum -y install redhat-lsb || exit 1
+yum -y install telnet sysstat nc bind-utils iputils
 yum -y install centos-release-scl ; yum install -y rh-python36 ; yum -y install python38 || yum -y install python36
 cd /root && tar -zxvf installer.tgz || exit 1
 cd aerospike-server-* && ./asinstall || exit 1
@@ -145,6 +148,7 @@ yum -y install iptables wget tcpdump which redhat-lsb-core initscripts binutils 
 yum -y install dnsutils || yum -y install bind-utils
 yum -y install python
 yum -y install initscripts || exit 1
+yum -y install telnet sysstat nc bind-utils iputils
 yum -y install centos-release-scl ; yum install -y rh-python36 ; yum -y install python38 || yum -y install python36
 cd /root && tar -zxvf installer.tgz || exit 1
 cd aerospike-server-* && ./asinstall || exit 1
@@ -158,6 +162,7 @@ apt-get -y install libcurl4 || exit 1
 apt-get -y install ldap-utils || exit 1
 apt-get -y install python3-setuptools || exit 1
 apt-get -y install iptables wget dnsutils tcpdump python net-tools vim binutils iproute2 python3 libcurl4-openssl-dev less || exit 1
+apt-get -y install dnsutils iputils-ping telnet netcat sysstat
 wget https://github.com/bestmethod/tcconfig-ubuntu-venv/archive/tc.tar.gz || exit 1
 tar -zxvf tc.tar.gz || exit 1
 mv /tcconfig-ubuntu-venv-tc /tcconfig || exit 1
@@ -257,6 +262,7 @@ yum -y install dnsutils || yum -y install bind-utils
 yum -y install python
 yum -y install initscripts || exit 1
 yum -y install redhat-lsb || exit 1
+yum -y install telnet sysstat nc bind-utils iputils
 yum -y install centos-release-scl ; yum install -y rh-python36 ; yum -y install python38 || yum -y install python36
 cd /root && tar -zxvf installer.tgz || exit 1
 cd aerospike-server-* && ./asinstall || exit 1