+++ CHANGELOG.md
@@ -5,9 +5,7 @@
 * support arm deployments
 
 #### 4.1.0 - TODO ITEMS
-* run script every time the node starts (crontab @reboot)
 * add option to deploy AMS (monitoring stack)
-* add install script to install basic linux tools (ip-utils, dns-utils, ping, telnet, netcat, netstat, ss, iostat, mpstat, sar)
 
 #### v4.0.0
 * add support for centos / rhel-based distros v 7 and 8 in aws
@@ -22,6 +20,8 @@
 * update readme and help pages
 * update scripts/
 * make help pages work without verifying if backend is working
+* install a script to run before aerospike starts and after it stops
+* add basic troubleshooting tools to templates
 
 #### v3.1.2
 * `make-cluster` and `cluster-grow`: automatically add `cluster-name` to `aerospike.conf` unless specified not to by the `-O` switch